+++ duckscript_sdk/src/sdk/std/collections/mod.rs
@@ -2,6 +2,8 @@ pub(crate) mod array;
 mod array_is_empty;
 pub(crate) mod array_length;
 mod range;
+mod read_properties;
+mod write_properties;
 
 use crate::utils::pckg;
 use duckscript::types::command::Commands;
@@ -16,6 +18,8 @@ pub(crate) fn load(commands: &mut Commands, parent: &str) -> Result<(), ScriptEr
     commands.set(array_is_empty::create(&package))?;
     commands.set(array_length::create(&package))?;
     commands.set(range::create(&package))?;
+    commands.set(read_properties::create(&package))?;
+    commands.set(write_properties::create(&package))?;
 
     Ok(())
 }