+++ duckscript_sdk/src/sdk/std/collections/read_properties/mod.rs
@@ -1,0 +1,64 @@
+use crate::utils::pckg;
+use duckscript::types::command::{Command, CommandResult, Commands};
+use duckscript::types::instruction::Instruction;
+use duckscript::types::runtime::StateValue;
+use java_properties::read;
+use std::collections::HashMap;
+
+#[cfg(test)]
+#[path = "./mod_test.rs"]
+mod mod_test;
+
+struct CommandImpl {
+    package: String,
+}
+
+impl Command for CommandImpl {
+    fn name(&self) -> String {
+        pckg::concat(&self.package, "ReadProperties")
+    }
+
+    fn aliases(&self) -> Vec<String> {
+        vec!["read_properties".to_string()]
+    }
+
+    fn help(&self) -> String {
+        include_str!("help.md").to_string()
+    }
+
+    fn requires_context(&self) -> bool {
+        true
+    }
+
+    fn run_with_context(
+        &self,
+        arguments: Vec<String>,
+        _state: &mut HashMap<String, StateValue>,
+        variables: &mut HashMap<String, String>,
+        _output_variable: Option<String>,
+        _instructions: &Vec<Instruction>,
+        _commands: &mut Commands,
+        _line: usize,
+    ) -> CommandResult {
+        if arguments.len() < 1 {
+            CommandResult::Error("Missing properties text argument.".to_string())
+        } else {
+            match read(arguments[0].as_bytes()) {
+                Ok(data) => {
+                    for (key, value) in &data {
+                        variables.insert(key.to_string(), value.to_string());
+                    }
+
+                    CommandResult::Continue(Some(data.len().to_string()))
+                }
+                Err(error) => CommandResult::Error(error.to_string()),
+            }
+        }
+    }
+}
+
+pub(crate) fn create(package: &str) -> Box<dyn Command> {
+    Box::new(CommandImpl {
+        package: package.to_string(),
+    })
+}